wip: move login logic to its own fragment
