+++ app/src/main/java/com/chesire/malime/view/login/ServiceSelectionFragment.kt
@@ -1,0 +1,6 @@
+package com.chesire.malime.view.login
+
+import android.support.v4.app.Fragment
+
+class ServiceSelectionFragment: Fragment() {
+}
